# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Git Bridge API",
    "description": "Multi-provider git repository mirroring service.\n\nMirrors repositories between CodeCommit, GitLab and GitHub. Sync is event-driven: CodeCommit events arrive over SQS (not an HTTP endpoint), while GitLab and GitHub push events arrive on the webhook endpoints below. A manual retry endpoint re-runs a sync for a single repository.\n\nAll endpoints here are machine-facing and reachable on the public host. Webhook senders are configured in the respective providers; the retry endpoint is called by operators.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Liveness probe",
        "description": "Returns service health status. Used by the Kubernetes liveness probe.",
        "responses": {
          "200": {
            "description": "Service is running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "status": "ok",
                  "service": "git-bridge"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Readiness probe",
        "description": "Returns service readiness status. Used by the Kubernetes readiness probe.",
        "responses": {
          "200": {
            "description": "Service is ready to accept requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "status": "ok",
                  "service": "git-bridge"
                }
              }
            }
          }
        }
      }
    },
    "/api-docs": {
      "get": {
        "tags": [
          "Docs"
        ],
        "summary": "API documentation",
        "description": "This Swagger UI page.",
        "responses": {
          "200": {
            "description": "HTML documentation page",
            "content": {
              "text/html": {}
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "tags": [
          "Docs"
        ],
        "summary": "OpenAPI specification",
        "description": "The source of this document. A unit test keeps it in sync with the route table.",
        "responses": {
          "200": {
            "description": "OpenAPI 3.0 document",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/webhook/gitlab": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "GitLab push event webhook",
        "description": "Receives push events from GitLab and triggers a mirror sync for the matching repository.\n\nThe `project.path_with_namespace` field is matched against `target_path` or `source_path` in the repo configuration.\n\nThe sync runs in the background, so a 200 means the event was accepted — not that the mirror finished.",
        "parameters": [
          {
            "name": "X-Gitlab-Token",
            "in": "header",
            "required": false,
            "description": "Secret token for webhook verification. Required only when `WEBHOOK_GITLAB_SECRET` is configured.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitLabPushEvent"
              },
              "example": {
                "event_name": "push",
                "ref": "refs/heads/main",
                "repository": {
                  "name": "my-repo"
                },
                "project": {
                  "path_with_namespace": "team/my-repo"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event accepted, mirror sync started asynchronously",
            "content": {
              "application/json": {
                "example": {
                  "status": "accepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "405": {
            "description": "Method not allowed (POST only)"
          }
        }
      }
    },
    "/webhook/github": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "GitHub push event webhook",
        "description": "Receives push events from GitHub and triggers a mirror sync for the matching repository.\n\nThe `repository.full_name` field is matched against `target_path` or `source_path` in the repo configuration.\n\nThe sync runs in the background, so a 200 means the event was accepted — not that the mirror finished.",
        "parameters": [
          {
            "name": "X-Hub-Signature-256",
            "in": "header",
            "required": false,
            "description": "HMAC-SHA256 signature of the request body (`sha256=<hex>`). Required only when `WEBHOOK_GITHUB_SECRET` is configured.",
            "schema": {
              "type": "string",
              "example": "sha256=abc123..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubPushEvent"
              },
              "example": {
                "ref": "refs/heads/main",
                "repository": {
                  "name": "my-repo",
                  "full_name": "org/my-repo"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event accepted, mirror sync started asynchronously",
            "content": {
              "application/json": {
                "example": {
                  "status": "accepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Invalid or missing secret"
          },
          "405": {
            "description": "Method not allowed (POST only)"
          }
        }
      }
    },
    "/retry/mirror": {
      "post": {
        "tags": [
          "Retry"
        ],
        "summary": "Manually re-run a mirror sync",
        "description": "Re-runs the mirror for one repository, for use after a failed sync.\n\nAuthentication is a bearer token compared in constant time. When `RETRY_API_TOKEN` is empty the endpoint is disabled and every request gets 404 — the route stays registered so the response is uniform whether or not the feature is on.\n\nThe sync runs in a background goroutine, so a 200 means the request was **accepted**, not that the mirror succeeded — and not even that the repository exists. An unknown `repo`, or a direction the repository's own `direction` disallows, still returns 200 here and only fails inside the goroutine. Check the pod logs (`source=retry-api`) or the Slack notification for the outcome.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "`Bearer <RETRY_API_TOKEN>`. The literal `Bearer ` prefix is required.",
            "schema": {
              "type": "string",
              "example": "Bearer s3cr3t-token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryRequest"
              },
              "example": {
                "repo": "git-bridge-test",
                "direction": "auto",
                "ref": "refs/heads/main"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retry accepted, mirror sync started asynchronously. Echoes back what was queued — this does not mean the repo exists or the sync will succeed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryAccepted"
                },
                "example": {
                  "status": "accepted",
                  "repo": "git-bridge-test",
                  "direction": "auto",
                  "ref": "refs/heads/main",
                  "queued_at": "2026-07-28T03:25:28Z"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, missing `repo`, or a `direction` outside the allowed values. An unknown repo is NOT rejected here — see the 200 description."
          },
          "401": {
            "description": "Missing, malformed, or incorrect bearer token"
          },
          "404": {
            "description": "Endpoint disabled (`RETRY_API_TOKEN` is empty)"
          },
          "405": {
            "description": "Method not allowed (POST only)"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "service": {
            "type": "string",
            "example": "git-bridge"
          }
        }
      },
      "GitLabPushEvent": {
        "type": "object",
        "properties": {
          "event_name": {
            "type": "string",
            "description": "Event type",
            "example": "push"
          },
          "ref": {
            "type": "string",
            "description": "Branch or tag reference",
            "example": "refs/heads/main"
          },
          "repository": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "my-repo"
              }
            }
          },
          "project": {
            "type": "object",
            "properties": {
              "path_with_namespace": {
                "type": "string",
                "description": "Full project path used for repo matching",
                "example": "team/my-repo"
              }
            }
          }
        }
      },
      "GitHubPushEvent": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string",
            "description": "Branch or tag reference",
            "example": "refs/heads/main"
          },
          "repository": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "my-repo"
              },
              "full_name": {
                "type": "string",
                "description": "Full repository name used for repo matching",
                "example": "org/my-repo"
              }
            }
          }
        }
      },
      "RetryRequest": {
        "type": "object",
        "required": [
          "repo"
        ],
        "properties": {
          "repo": {
            "type": "string",
            "description": "Repository name as declared under `repos[].name` in the config.",
            "example": "git-bridge-test"
          },
          "direction": {
            "type": "string",
            "description": "Mirror direction to run. Defaults to `auto` when omitted or empty. Matching is case-insensitive.\n\n`auto` resolves in this order: the repository's explicit `retry_direction` if set, otherwise a built-in fallback by the repository's `direction` (bidirectional resolves to target-to-source; one-way repositories resolve to their single allowed direction).",
            "enum": [
              "auto",
              "source-to-target",
              "target-to-source"
            ],
            "default": "auto",
            "example": "auto"
          },
          "ref": {
            "type": "string",
            "description": "Optional single ref to sync. Omit to sync every ref.",
            "example": "refs/heads/main"
          },
          "source": {
            "type": "string",
            "description": "Optional caller identity, recorded as the event's trigger. Only `cron` is accepted — the reconcile CronJob sets it so its scheduled calls are distinguishable from a hand-run retry, which is recorded as `retry-api`. Any other value is rejected with 400, keeping the trigger vocabulary closed.",
            "enum": [
              "cron"
            ],
            "example": "cron"
          },
          "dest": {
            "type": "string",
            "description": "The destination tip being overwritten. Required with `force`, and becomes the push's lease — without it the force would run against whatever the destination holds when the push finally executes, which after a minute-long fetch is not necessarily the commit the caller decided about. A force with no `dest` returns 400.",
            "example": "95a8c3753195f3e50fea456e7bea4c376737655e"
          },
          "force": {
            "type": "boolean",
            "description": "Apply a rewind the push guard would otherwise withhold.\n\nThe mirror refuses a push that would move a ref backwards at the destination — the destination already contains what this side holds, so the write would discard commits there rather than mirror them. That refusal also catches a rewind somebody meant to make, and this flag is how they say which it is.\n\nIt does NOT skip the lease: the push still carries the destination tip it was checked against, so a commit arriving between the check and the write is still refused. Three limits keep it narrow — `ref` is required, so one call moves at most one ref; `dest` is required, so it can only discard the commit the caller looked at; and `source: cron` is refused, so the unattended reconcile can never carry a bypass. All three violations return 400.",
            "default": false,
            "example": true
          },
          "actor": {
            "type": "string",
            "description": "Optional attribution recorded on the history event and included in the Slack message. The console button fills it from the portal's `X-Auth-User`.",
            "example": "alice"
          }
        }
      },
      "RetryAccepted": {
        "type": "object",
        "description": "Echo of the queued retry. Returned before the sync runs.",
        "properties": {
          "status": {
            "type": "string",
            "example": "accepted"
          },
          "repo": {
            "type": "string",
            "example": "git-bridge-test"
          },
          "direction": {
            "type": "string",
            "description": "The direction as received — `auto` is echoed literally, not the resolved value.",
            "example": "auto"
          },
          "ref": {
            "type": "string",
            "description": "Empty when no ref was requested.",
            "example": "refs/heads/main"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-28T03:25:28Z"
          }
        }
      },
      "CodeCommitEvent": {
        "type": "object",
        "description": "EventBridge event consumed via SQS (not an HTTP endpoint)",
        "properties": {
          "detail": {
            "type": "object",
            "properties": {
              "repositoryName": {
                "type": "string",
                "description": "CodeCommit repository name",
                "example": "my-repo"
              },
              "referenceName": {
                "type": "string",
                "description": "Branch or tag reference",
                "example": "refs/heads/main"
              },
              "referenceType": {
                "type": "string",
                "enum": [
                  "branch",
                  "tag"
                ],
                "example": "branch"
              },
              "event": {
                "type": "string",
                "enum": [
                  "referenceCreated",
                  "referenceUpdated",
                  "referenceDeleted"
                ],
                "example": "referenceUpdated"
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "Kubernetes health check endpoints"
    },
    {
      "name": "Docs",
      "description": "API documentation"
    },
    {
      "name": "Webhooks",
      "description": "Git provider webhook endpoints for mirror sync"
    },
    {
      "name": "Retry",
      "description": "Manual mirror retry"
    }
  ]
}